avancéAngular

+++ src/app/cour/course-delete/course-delete.component.ts
@@ -8,7 +8,7 @@ import { CourseService } from 'src/app/service/course.service';
 })
 export class CourseDeleteComponent implements OnInit {
   constructor(private CourseService:CourseService){}
-  search=""
+  search="";
   course!:any
   idc!:String
   listeCourse!:course[]
@@ -24,17 +24,28 @@ export class CourseDeleteComponent implements OnInit {
   
   }
   delete(id: string) {
-    this.CourseService.deleteCourse(id).subscribe(
-      () => {
-        console.log(`La course avec l'ID ${id} a été supprimée avec succès.`);
-        // Actualiser la page après la suppression
-        location.reload();
-      },
-      (error) => {
-        console.error(`Erreur lors de la suppression de la course avec l'ID ${id} :`, error);
-      }
-    );
+    // Afficher la boîte de dialogue de confirmation
+    const confirmed = window.confirm('Voulez-vous vraiment supprimer cette course ?');
+    
+    // Si l'utilisateur clique sur "Oui" dans la boîte de dialogue
+    if (confirmed) {
+      // Supprimer la course
+      this.CourseService.deleteCourse(id).subscribe(
+        () => {
+          console.log(`La course avec l'ID ${id} a été supprimée avec succès.`);
+          // Actualiser la page après la suppression
+          location.reload();
+        },
+        (error) => {
+          console.error(`Erreur lors de la suppression de la course avec l'ID ${id} :`, error);
+        }
+      );
+    } else {
+      // Si l'utilisateur clique sur "Non" dans la boîte de dialogue
+      console.log('Suppression de la course annulée.');
+    }
   }
+  
   modifier(id: string){
     this.idc=id;
     console.log(id);
@@ -51,5 +62,9 @@ trier(){
 }
   afficher(){
     console.log(this.course);
+  }
+
+  getphoto(photo :string){
+    return this.CourseService.getPhoto(photo);
   }
 }
